+++ src/components/AdsSection.tsx
@@ -1,4 +1,4 @@
-import siteData from "@/data.json";
+import siteData from '@/data/home.json';
 import Button from "./UI/Button";
 import Section from "./UI/Section";
