+++ src/components/Customers.tsx
@@ -1,6 +1,6 @@
 'use client';
 import Image from "next/image";
-import siteData from "@/data.json";
+import siteData from '@/data/home.json';
 import Section from "./UI/Section";
 import Slider from "react-slick";
