+++ src/data/ads.json
@@ -1,0 +1,8 @@
+{
+    "title": "Our Other services",
+    "header1": "Discover Your Next Treasure: Browse Our Ads Listings Today!",
+    "paragraph": "Welcome to our vibrant marketplace, where opportunities abound and treasures await discovery! Whether you&#39;re hunting for that perfect item, seeking to sell your goods, or simply exploring what&#39;s out there, you&#39;ve come to the right place. Dive into our extensive collection of ads listings and uncover a world of possibilities. From electronics to fashion, real estate to services, our platform connects buyers and sellers, making transactions seamless and exciting. Join our community today and embark on your journey to find or showcase that special something.",
+
+    "grandTitle": "Our Other Services",
+    "grandSubTitle": "Services you may be intersted in"
+}
